Move callback injecting logic to the class, use weakset for handling cache of already processed agents
