+++ sdks/python/src/opik/integrations/adk/__init__.py
@@ -1,4 +1,4 @@
 from .opik_tracer import OpikTracer
-from .callback_injector import track_adk_agent_recursive
+from .recursive_callback_injector import track_adk_agent_recursive
 
 __all__ = ["OpikTracer", "track_adk_agent_recursive"]
