+++ sdks/python/src/opik/integrations/adk/recursive_callback_injector.py
@@ -1,0 +1,120 @@
+from typing import TypeVar
+from . import opik_tracer
+import logging
+import weakref
+from opik import _logging
+
+from google.adk.tools import agent_tool
+from google.adk import agents
+
+LOGGER = logging.getLogger(__name__)
+
+ADKAgent = TypeVar("ADKAgent", bound=agents.BaseAgent)
+
+
+class RecursiveCallbackInjector:
+    def __init__(self) -> None:
+        self._tracked_agents: weakref.WeakSet[agents.BaseAgent] = weakref.WeakSet()
+
+    def inject(
+        self,
+        root_agent: agents.BaseAgent,
+        tracer: opik_tracer.OpikTracer,
+    ) -> None:
+        self._process_agent(root_agent, tracer)
+
+    def _add_callbacks_to_agent(
+        self, agent: agents.BaseAgent, tracer: opik_tracer.OpikTracer
+    ) -> None:
+        callback_fields = {
+            "before_agent_callback": tracer.before_agent_callback,
+            "after_agent_callback": tracer.after_agent_callback,
+            "before_model_callback": tracer.before_model_callback,
+            "after_model_callback": tracer.after_model_callback,
+            "before_tool_callback": tracer.before_tool_callback,
+            "after_tool_callback": tracer.after_tool_callback,
+        }
+
+        for callback_field_name, callback_func in callback_fields.items():
+            if not hasattr(agent, callback_field_name):
+                continue
+
+            current_callback_value = getattr(agent, callback_field_name)
+            if current_callback_value is None:
+                setattr(agent, callback_field_name, callback_func)
+            elif isinstance(current_callback_value, list):
+                current_callback_value.append(callback_func)
+            else:
+                setattr(
+                    agent, callback_field_name, [current_callback_value, callback_func]
+                )
+
+    def _process_agent(
+        self,
+        agent: agents.BaseAgent,
+        tracer: opik_tracer.OpikTracer,
+    ) -> None:
+        if agent in self._tracked_agents:
+            return
+
+        self._add_callbacks_to_agent(agent, tracer)
+        self._process_sub_agents(agent, tracer)
+        self._process_tools(agent, tracer)
+
+        self._tracked_agents.add(agent)
+
+    def _process_sub_agents(
+        self,
+        agent: agents.BaseAgent,
+        tracer: opik_tracer.OpikTracer,
+    ) -> None:
+        if not hasattr(agent, "sub_agents"):
+            return
+
+        for sub_agent in agent.sub_agents:
+            try:
+                self._process_agent(sub_agent, tracer)
+            except Exception as e:
+                LOGGER.warning(f"Failed to track subagent: {e}")
+
+    def _process_tools(
+        self,
+        agent: agents.BaseAgent,
+        tracer: opik_tracer.OpikTracer,
+    ) -> None:
+        if not hasattr(agent, "tools"):
+            return
+
+        for tool in agent.tools:
+            if not isinstance(tool, agent_tool.AgentTool):
+                continue
+            try:
+                self._process_agent(tool.agent, tracer)
+            except Exception as e:
+                LOGGER.warning(f"Failed to track agent tool: {e}")
+
+
+_recursive_callback_injector = RecursiveCallbackInjector()
+
+
+def track_adk_agent_recursive(
+    root_agent: ADKAgent, tracer: opik_tracer.OpikTracer
+) -> ADKAgent:
+    """
+    Recursively adds opik tracer callbacks to the agent, its subagents, and agent tools.
+
+    Args:
+        root_agent: The root ADK agent to track
+        tracer: The OpikTracer instance to use for tracking
+
+    Returns:
+        The modified root agent with tracking enabled
+    """
+    _logging.log_once_at_level(
+        logging.INFO,
+        "`track_adk_agent_recursive` is experimental feature. Please let us know if something is not working as expected: https://github.com/comet-ml/opik/issues",
+        logger=LOGGER,
+    )
+    _recursive_callback_injector.inject(root_agent, tracer)
+
+    return root_agent
